requirements.txt
telethon
langdetect
openai
flask
python-dotenv
apscheduler
